game is fully functional, inventory decrease functionality is now accurate, customers making choices are fixed, removed do while loop of customer purchases and changed to a controlled while loop, difficulty significantly harder until more customer objects are created

+++ Inventory.cs
@@ -9,7 +9,7 @@ namespace LemonadeStand
     class Inventory
     {
         //member variables
-        public double quantity;
+        public int quantity;
         //ctor
         public Inventory()
         {
